Removed export window
