+++ Editor/CustomEditors/UrdfRobotEditor.cs
@@ -12,6 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */  
 
+using System.IO;
 using UnityEditor;
 using UnityEngine;
 
@@ -22,6 +23,7 @@ namespace RosSharp.Urdf.Editor
     {
         private UrdfRobot urdfRobot;
         private static GUIStyle buttonStyle;
+        private string exportRoot = "";
         SerializedProperty axisType;
 
         public void OnEnable()
@@ -34,6 +36,10 @@ namespace RosSharp.Urdf.Editor
                 buttonStyle = new GUIStyle(EditorStyles.miniButtonRight) { fixedWidth = 75 };
 
             urdfRobot = (UrdfRobot) target;
+
+            EditorGUILayout.PropertyField(axisType, new GUIContent("Axis Type"));
+            serializedObject.ApplyModifiedProperties();
+            UrdfRobotExtensions.CorrectAxis(urdfRobot.gameObject);
 
             GUILayout.Space(5);
             GUILayout.Label("All Rigidbodies", EditorStyles.boldLabel);
@@ -55,22 +61,32 @@ namespace RosSharp.Urdf.Editor
             EditorGUILayout.EndHorizontal();
 
             GUILayout.Space(5);
-            EditorGUILayout.PropertyField(axisType, new GUIContent("Axis Type"));
-            serializedObject.ApplyModifiedProperties();
-            UrdfRobotExtensions.CorrectAxis(urdfRobot.gameObject);
+            
             GUILayout.Label("Helper Scripts", EditorStyles.boldLabel);
             DisplaySettingsToggle(new GUIContent("Controller Script"), urdfRobot.AddController);
             DisplaySettingsToggle(new GUIContent("Forward Kinematics Script"), urdfRobot.AddFkRobot);
 
             GUILayout.Space(5);
-            if (GUILayout.Button("Export robot to URDF file"))
+
+            GUILayout.Label("URDF Options", EditorStyles.boldLabel);
+            GUILayout.BeginHorizontal();
+            StlWriter.fileType = (StlWriter.FileType) EditorGUILayout.EnumPopup("Export new URDF meshes to:", StlWriter.fileType);
+            GUILayout.Label("STL files");
+            bool export = GUILayout.Button("Export to URDF");
+            EditorGUILayout.EndHorizontal();
+
+            //Export Robot button
+            if (export)
             {
-                // Get existing open window or if none, make a new one:
-                UrdfExportEditorWindow window = (UrdfExportEditorWindow)EditorWindow.GetWindow(typeof(UrdfExportEditorWindow));
-                window.urdfRobot = urdfRobot;
-                window.minSize = new Vector2(500, 200);
-                window.GetEditorPrefs();
-                window.Show();
+                exportRoot = EditorUtility.OpenFolderPanel("Select export directory", exportRoot, "");
+
+                if (exportRoot == "" || !Directory.Exists(exportRoot))
+                    EditorUtility.DisplayDialog("Export Error", "Export root folder must be defined and folder must exist.", "Ok");
+                else
+                {
+                    urdfRobot.ExportRobotToUrdf(exportRoot, exportRoot);
+                    SetEditorPrefs();
+                }
             }
 
             GUILayout.Space(5);
@@ -94,6 +110,11 @@ namespace RosSharp.Urdf.Editor
             if (GUILayout.Button("Disable", buttonStyle))
                 handler(false);
             EditorGUILayout.EndHorizontal();
+        }
+
+        private void SetEditorPrefs()
+        {
+            EditorPrefs.SetString("UrdfExportRoot", exportRoot);
         }
 
     }
