--- Editor/MenuItems/UrdfExportEditorWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fe4155eae653842448a6770a238b08a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
